More canvas refactoring.
