+++ src/main/java/org/geppetto/frontend/messages/OutboundMessages.java
@@ -13,7 +13,6 @@ public enum OutboundMessages {
 	SERVER_AVAILABLE("server_available"),
 	EXPERIMENT_RUNNING("experiment_running"),
 	INFO_MESSAGE("info_message"),
-	RELOAD_CANVAS("reload_canvas"),
 	SIMULATION_CONFIGURATION("simulation_configuration"),
 	ERROR("generic_error"),
 	ERROR_RUNNING_EXPERIMENT("error_running_experiment"),
