+++ src/main/webapp/js/geppettoProject/ExperimentsController.js
@@ -267,7 +267,7 @@ define(function (require) {
                     }
 
                 } else {
-                    GEPPETTO.FE.infoDialog(GEPPETTO.Resources.CANT_PLAY_EXPERIMENT, "Experiment " + experiment.name + " with id " + experiment.id + " isn't completed, and can't be played.");
+                    GEPPETTO.ModalFactory.infoDialog(GEPPETTO.Resources.CANT_PLAY_EXPERIMENT, "Experiment " + experiment.name + " with id " + experiment.id + " isn't completed, and can't be played.");
                 }
             },
             
@@ -360,7 +360,7 @@ define(function (require) {
                     this.worker = new Worker("geppetto/js/geppettoProject/ExperimentWorker.js");
 
                     // tells worker to update each half a second
-                    this.worker.postMessage([GEPPETTO.Events.Experiment_play, GEPPETTO.getVARS().playTimerStep, this.playOptions.step]);
+                    this.worker.postMessage([GEPPETTO.Events.Experiment_play, GEPPETTO.ExperimentsController.playTimerStep, this.playOptions.step]);
 
                     var that = this;
                     // receives message from web worker
