+++ src/main/webapp/js/pages/geppetto/GEPPETTO.Main.js
@@ -1,6 +1,5 @@
 /**
- * Main class for handling user interface evens associated with: Simulation Controls,
- * alert & info messages, and server side communication
+ * Geppetto entry point
  *
  * @author matteo@openworm.org (Matteo Cantarelli)
  * @author giovanni@openworm.org (Giovanni Idili)
@@ -9,11 +8,11 @@
 define(function (require) {
     return function (GEPPETTO) {
 
-        var $ = require('jquery'),
-            React = require('react'),
-            InfoModal = require('../../components/controls/modals/InfoModal'),
-            ProjectNode = require('../../geppettoProject/model/ProjectNode'),
-            ReactDOM = require('react-dom');
+        var $ = require('jquery');
+        var React = require('react');
+        var InfoModal = require('../../components/controls/modals/InfoModal');
+        var ProjectNode = require('../../geppettoProject/model/ProjectNode');
+        var ReactDOM = require('react-dom');
 
         /**
          * @class GEPPETTO.Main
@@ -22,14 +21,55 @@ define(function (require) {
 
             idleTime: 0,
             disconnected: false,
-            status: 0,
-            statusWorker : null,
-            localStorageEnabled : false,
+            statusWorker: null,
+            localStorageEnabled: false,
 
+            /**
+             *
+             */
+            createChannel: function () {
+                // Change link from blank to self for embedded environments
+                if (window.EMBEDDED && window.EMBEDDEDURL !== "/" && typeof handleRequest == 'undefined') {
+                    handleRequest = function (e) {
+                        if (window.EMBEDDEDURL.indexOf(e.origin) != -1) {
+                            if (e.data.command == 'loadSimulation') {
+                                if (e.data.projectId) {
+                                    GEPPETTO.Console.executeCommand('Project.loadFromID(' + e.data.projectId + ')');
+                                }
+                                else if (e.data.url) {
+                                    GEPPETTO.Console.executeCommand('Project.loadFromURL("' + e.data.url + '")');
+                                }
+                            }
+                            else if (e.data.command == 'removeWidgets') {
+                                GEPPETTO.Console.executeCommand('G.removeWidget()');
+                            }
+                            else {
+                                eval(e.data.command);
+                            }
+                        }
+                    };
+                    // we have to listen for 'message'
+                    window.addEventListener('message', handleRequest, false);
+                    if ($.isArray(window.EMBEDDEDURL)) {
+                        window.parent.postMessage({"command": "ready"}, window.EMBEDDEDURL[0]);
+                    }
+                    else {
+                        window.parent.postMessage({"command": "ready"}, window.EMBEDDEDURL);
+                    }
+                }
+            },
+
+            /**
+             *
+             * @returns {null}
+             */
             getStatusWorker: function () {
                 return this.statusWorker;
             },
 
+            /**
+             *
+             */
             startStatusWorker: function () {
                 //create web worker for checking status
                 if (this.statusWorker != undefined) {
@@ -64,6 +104,8 @@ define(function (require) {
              */
             init: function () {
                 GEPPETTO.MessageSocket.connect(GEPPETTO.MessageSocket.protocol + window.location.host + '/' + window.BUNDLE_CONTEXT_PATH + '/GeppettoServlet');
+                GEPPETTO.Events.listen();
+                this.createChannel();
                 GEPPETTO.Console.debugLog(GEPPETTO.Resources.GEPPETTO_INITIALIZED);
             },
 
@@ -78,8 +120,9 @@ define(function (require) {
                         //first time check, asks if user is still there
                         if (GEPPETTO.Main.idleTime > allowedTime) { // 5 minutes
 
+                            //TODO Matteo: Make a function to create a custom Info modal inside ModalFactory and use it from here.
                             var infoFactory = React.createFactory(InfoModal);
-                            ReactDOM.render(infoFactory({ show: true, keyboard: false }), document.getElementById('modal-region'));
+                            ReactDOM.render(infoFactory({show: true, keyboard: false}), document.getElementById('modal-region'));
 
                             $('#infomodal-title').html("Zzz");
                             $('#infomodal-text').html(GEPPETTO.Resources.IDLE_MESSAGE);
@@ -97,6 +140,7 @@ define(function (require) {
                         //second check, user isn't there or didn't click yes, disconnect
                         if (GEPPETTO.Main.idleTime > timeOut) {
 
+                            //TODO Matteo: Make a function to create a custom Info modal inside ModalFactory and use it from here.
                             var infoFactory = React.createFactory(InfoModal);
                             ReactDOM.render(infoFactory({
                                 show: true,
@@ -110,16 +154,9 @@ define(function (require) {
 
                             GEPPETTO.Main.idleTime = 0;
                             GEPPETTO.Main.disconnected = true;
-                            GEPPETTO.FE.disableSimulationControls();
                             GEPPETTO.MessageSocket.close();
 
-                            var webGLStarted = GEPPETTO.init(GEPPETTO.FE.createContainer());
-                            var webWorkersSupported = (typeof (Worker) !== "undefined") ? true : false;
 
-                            if (!webGLStarted || !webWorkersSupported) {
-                                GEPPETTO.FE.notifyInitErrors(webGLStarted, webWorkersSupported);
-                                GEPPETTO.MessageSocket.close();
-                            }
                         }
                     }
                 }
@@ -127,37 +164,36 @@ define(function (require) {
 
         };
 
-        // ============================================================================
-        // Application logic.
-        // ============================================================================
-
         $(document).ready(function () {
+            GEPPETTO.Console.createConsole();
             var webWorkersSupported = (typeof (Worker) !== "undefined") ? true : false;
 
             //make sure webgl started correctly
             if (!webWorkersSupported) {
-                GEPPETTO.FE.notifyInitErrors(webGLStarted, webWorkersSupported);
+                GEPPETTO.Console.debugLog(GEPPETTO.Resources.WORKERS_NOT_SUPPORTED);
+                GEPPETTO.ModalFactory.infoDialog(GEPPETTO.Resources.WORKERS_NOT_SUPPORTED, GEPPETTO.Resources.WORKERS_NOT_SUPPORTED_MESSAGE);
             } else {
-                GEPPETTO.FE.initialEvents();
 
                 //Increment the idle time counter every minute.
                 setInterval(GEPPETTO.Main.idleCheck, 240000); // 1 minute
                 var here = $(this);
+
                 //Zero the idle timer on mouse movement.
                 here.mousemove(function (e) {
                     if (GEPPETTO.Main.idleTime > -1) {
                         GEPPETTO.Main.idleTime = 0;
                     }
                 });
+
                 here.keypress(function (e) {
                     if (GEPPETTO.Main.idleTime > -1) {
                         GEPPETTO.Main.idleTime = 0;
                     }
                 });
 
-                //Initialize websocket functionality
                 GEPPETTO.Main.init();
 
+                //TODO Matteo: All the code below needs to be removed creating a component for the tabbed UI
                 var visibleExperiments = false;
                 $('#experimentsButton').click(function (e) {
                     if (!visibleExperiments) {
